[ALS-8297] Move loadResources (#478)
The loadResources method can't be in the higher layout because the user
doesn't have a token yet for the resources api call that happens when
the federated feature flag is on. This causes the application to not
allow login. Moving it to before the calls for stats and results fixes
this issue because they only load on pages that a logged in user
accesses.